scalability test

+++ sim/shared_lib/bgp.h
@@ -92,8 +92,8 @@ int get_route_size(route_t *r);
 int write_route_msg(uint8_t *route, route_t *input);
 void generate_bgp_msg(bgp_msg_t **pp_bgp_msg, route_t *input, uint32_t asn, uint8_t oprt_type);
 route_node_t* get_selected_route_node(route_node_t *p_rns);
-void add_route(route_node_t **pp_rns, uint32_t src_asn, route_t *src_route, uint32_t *import_policy);
-void del_route(route_node_t **pp_rns, uint32_t src_asn, route_t *src_route, uint32_t *import_policy, route_node_t *p_old_best_rn);
-void execute_export_policy(rs_inner_msg_t **pp_inner_msgs, uint32_t num, uint32_t *export_policy, uint32_t src_asn, uint32_t src_next_hop, uint8_t oprt_type, route_t *src_route);
+void add_route(route_node_t **pp_rns, uint32_t src_asn, route_t *src_route, uint8_t *import_policy);
+void del_route(route_node_t **pp_rns, uint32_t src_asn, route_t *src_route, uint8_t *import_policy, route_node_t *p_old_best_rn);
+void execute_export_policy(rs_inner_msg_t **pp_inner_msgs, uint32_t num, uint8_t *export_policy, uint32_t src_asn, uint32_t src_next_hop, uint8_t oprt_type, route_t *src_route);
 
 #endif
